On branch 17.0.1  Changes to be committed: 	modified:   README.md 	modified:   odoo_install.sh 	deleted:    odoo_install_debian.sh
Added options for remote PostgreSQL setup.

Added a check to grab Ubuntu version for script
adjustments.

Set WKHTMLTOPDF links to 0.12.6-1 for Ubuntu 20.04

Added the installation of software-properties-common
in order to make add-apt-repository work on
fresh Ubuntu installs.

Added a check for which Ubuntu version was running
and then add the correct main repository instead of
xenial main.

Force pyOpenSSL version to 22.1.0 due to issues
when installing as Enterprise on 22.04.